Add files via upload
This project focuses on the analysis, simulation, and layout design of photovoltaic (PV) panels installed on a residential rooftop. The main objective is to evaluate the electrical performance of the selected solar module and to determine the optimal number and arrangement of panels under real climatic conditions.
The simulation process was carried out using TRNSYS to model the I–V and P–V characteristics and to analyze the annual variation of irradiance, temperature, and power output. Subsequently, PVsyst was employed to design the 3D layout of the rooftop, assess shading effects, and select appropriate inverters for each string configuration. PVsyst was also used to calculate the total annual electricity generation of the system.